Move updateMenteeData endpoint to Program Controller (#268)
