+++ src/main/java/org/sefglobal/scholarx/controller/MentorController.java
@@ -68,18 +68,6 @@ public class MentorController {
     return mentorService.getAllMenteesOfMentor(id, state);
   }
 
-  @PutMapping("/{id}/mentee")
-  @ResponseStatus(HttpStatus.OK)
-  public Mentee updateMenteeData(
-    @PathVariable long id,
-    Authentication authentication,
-    @Valid @RequestBody Mentee mentee
-  )
-    throws ResourceNotFoundException, BadRequestException {
-    Profile profile = (Profile) authentication.getPrincipal();
-    return mentorService.updateMenteeData(profile.getId(), id, mentee);
-  }
-
   @GetMapping("/{id}/mentee")
   @ResponseStatus(HttpStatus.OK)
   public Mentee getLoggedInMentee(
