+++ src/main/java/org/sefglobal/scholarx/service/ProgramService.java
@@ -378,4 +378,42 @@ public class ProgramService {
         log.error(msg);
         throw new NoContentException(msg);
     }
+
+    /**
+     * Update the application of a {@link Mentee}
+     *
+     * @param profileId which is the Profile id of the {@link Mentee} to be updated
+     * @param programId  which is the Program id of the {@link Mentee} to be updated
+     * @param mentee    with the application of the mentee to be updated
+     * @return the updated {@link Mentee}
+     *
+     * @throws ResourceNotFoundException is thrown if the {@link Mentee} doesn't exist
+     * @throws BadRequestException       if the {@link Mentee} is not in the valid state
+     */
+    public Mentee updateMenteeData(long profileId, long programId, Mentee mentee)
+            throws ResourceNotFoundException, BadRequestException {
+        Optional<Mentee> optionalMentee = menteeRepository.findByProgramIdAndProfileId(programId, profileId);
+        if (!optionalMentee.isPresent()) {
+            String msg = "Error, Mentee by profile id: " + profileId + " and " +
+                    "program id: " + programId + " cannot be updated. " +
+                    "Mentee doesn't exist.";
+            log.error(msg);
+            throw new ResourceNotFoundException(msg);
+        }
+
+        Mentee existingMentee = optionalMentee.get();
+        if (EnrolmentState.PENDING.equals(existingMentee.getState())) {
+            existingMentee.setAppliedMentor(mentee.getAppliedMentor());
+            existingMentee.setIntent(mentee.getIntent());
+            existingMentee.setUniversity(mentee.getUniversity());
+            existingMentee.setYear(mentee.getYear());
+            existingMentee.setReasonForChoice(mentee.getReasonForChoice());
+        } else {
+            String msg = "Error, Application cannot be updated. " +
+                    "Mentee is not in a valid state.";
+            log.error(msg);
+            throw new BadRequestException(msg);
+        }
+        return menteeRepository.save(existingMentee);
+    }
 }
